+++ src/main/java/com/example/justpost/domain/post/DownloadType.java
@@ -1,4 +1,4 @@
-package com.example.justpost.domain;
+package com.example.justpost.domain.post;
 
 public enum DownloadType {
     GS_POST, CU_POST, CJ_POST,
