+++ src/main/java/com/example/justpost/domain/store/afterPost/NaverAfterPostConverter.java
@@ -1,6 +1,6 @@
 package com.example.justpost.domain.store.afterPost;
 
-import com.example.justpost.domain.InvoiceMap;
+import com.example.justpost.domain.post.InvoiceMap;
 import com.example.justpost.domain.utils.ExcelUtil;
 import com.example.justpost.domain.utils.FileUtil;
 import org.apache.poi.hssf.usermodel.HSSFWorkbook;
