+++ src/main/java/com/example/justpost/domain/store/post/CoupangPostConverter.java
@@ -1,8 +1,8 @@
 package com.example.justpost.domain.store.post;
 
-import com.example.justpost.domain.OrderColumnIndex;
+import com.example.justpost.domain.post.OrderColumnIndex;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
-import org.thymeleaf.util.StringUtils;
 
 import static com.example.justpost.domain.utils.StringUtil.getIndex;
