+++ src/main/java/com/example/justpost/domain/store/post/NaverPostConverter.java
@@ -1,6 +1,6 @@
 package com.example.justpost.domain.store.post;
 
-import com.example.justpost.domain.OrderColumnIndex;
+import com.example.justpost.domain.post.OrderColumnIndex;
 import com.example.justpost.domain.utils.ExcelUtil;
 import org.apache.poi.poifs.crypt.Decryptor;
 import org.apache.poi.poifs.crypt.EncryptionInfo;
