+++ src/main/java/com/example/justpost/service/post/PostService.java
@@ -1,9 +1,9 @@
 package com.example.justpost.service.post;
 
-import com.example.justpost.domain.ConvertType;
-import com.example.justpost.domain.DownloadType;
-import com.example.justpost.domain.Posts;
-import com.example.justpost.domain.post.*;
+import com.example.justpost.domain.post.ConvertType;
+import com.example.justpost.domain.post.DownloadType;
+import com.example.justpost.domain.post.Posts;
+import com.example.justpost.domain.postClient.*;
 import com.example.justpost.domain.store.post.PostConverter;
 import com.example.justpost.domain.store.post.PostConverterFactory;
 import com.example.justpost.domain.utils.FileUtil;
